requirements.txt
flask==2.3.3
rembg==2.0.66
Pillow==10.0.0
flask-cors==4.0.0
gunicorn==21.2.0